+++ CHANGELOG.md
@@ -42,6 +42,19 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ---
 
+## [v1.19.1](https://github.com/provenance-io/provenance/releases/tag/v1.19.1) - 2024-07-25
+
+### Bug Fixes
+
+* Use the correct HRP in the CLI queries [#2108](https://github.com/provenance-io/provenance/issues/2108).
+
+### Full Commit History
+
+* https://github.com/provenance-io/provenance/compare/v1.19.0...v1.19.1
+* https://github.com/provenance-io/provenance/compare/v1.18.0...v1.19.1
+
+---
+
 ## [v1.19.0](https://github.com/provenance-io/provenance/releases/tag/v1.19.0) - 2024-07-19
 
 ### Features